Introducing separate calibration YAML file: Testing for one tract
Until now, all steps of single frame processing and calibration were
lumped together in a single YAML file. However, by introducing changes
to "DRP.yaml" within the "obs_vista" package, we now have distinct
steps for single frame processing (step1) and Calibration (step2)
before moving on to the coadd step. This necessitates the addition of
a new bps yaml file for step2, encompassing visit-level, tract-level,
and collection-level tasks.

With this commit, I have added a new YAML file named
"bpsAllCalibrate.yaml" and a corresponding slurm file named
"3.0_startCalibrate.slurm". Initial testing has been conducted for one
tract and with one node on CSD3.

It is important to note that users will need to modify the complete
butler address in the YAML file, as well as the address of SingleFrame
processing data, specified as an inCollection of step2.