Modify multi-visit YAML file: Executing step 3b
Until now, a multivisit YAML file existed alongside its corresponding
slurm files. However, as detailed in commit 5b0732a, adjustments were
required to align with recent updates to the LSST pipelines.

With this commit, I have updated the multivisit YAML file to execute
step3b of DRP.yaml, which is the last step of processing. Initial
testing has been conducted for one tract and with one node on CSD3.

Please, remember to update the complete butler address in the YAML
file, as well as the address of videoCoaddDetect, specified as an
inCollection of step3b.