Initialize pipeline run on LSST DP1 (ComCam) data
Until now, the pipeline had only been tested with HSC data. With the
availability of LSST ComCam DP1 data, we decided to test the pipeline
on one field, ECDFS, which also has VIDEO coverage in the CDFS region.

In this commit, we began a local run of the pipeline on example CDFS
data and will incrementally apply the necessary changes for this new
run. In this step, we creted skyMap of ComCam.