Generating list of reference catalogs and ingesting into Butler
Until now, we cross-matched the Monster reference catalogs with
VISTA-VIDEO-CDFS-DR5 and added VISTA fluxes to the refcats.

In this commit, we generated a list of HTM7 reference catalogs and
ingested them into Butler. Additionally, VISTA column names were
updated to ks_flux, j_flux, h_flux, z2_flux, and y2_flux to match the
naming convention of previous refcats in the pipeline.