Generate metadata for ComCam data ingestion
Until now, we successfully ran the pipeline on example CDFS VISTA data
through the coaddition step.

Before starting the final stage (measuring and forcedPhotometry), we
need to ingest deepCoadd images and detection catalogs from ComCam
data in ECDFS for the relevant tract and patch, in order to enable
forcedPhotometry across both NIR and optical bands.

In this commit, I added a Jupyter notebook that generates the required
metadata for deepCoadd images from ComCam.