Add script to filter VIDEO_CDFS images for ingestion
Until now, we transferred raw and confidence data from CSD3 to
Iridis6, including the full set of VIDEO images for both XMM and CDFS
fields at:

  /iridisfs/lsstphasec/data/private/VISTA/VIDEO/

In this commit, I added a Python script that identifies the VIDEO_CDFS
images by checking their headers and selecting only those within the
max/min bounds of the CDFS area. This ensures that only relevant
images are ingested for the CDFS pipeline run.