Add metadata and ingestion for ComCam detection catalogs
Until now, we generated metadata for ComCam deepCoadd images to
prepare for forcedPhotometry on ECDFS data. However, the associated
detection catalogs were missing. These catalogs are not available on
the RSP, since the DM team considers them unnecessary intermediate
products.

In this commit, I added metadata for the object_detection catalogs
(previously named deepCoadd_det) and ingested both the deepCoadd
images and the detection catalogs into the Butler. For DP1 we obtained
these files from the DM team, but for future data releases (DP2 and
DR1) we will need to update the pipeline to produce them ourselves.