Add Slurm file for ingesting VIRCAM CDFS data on Iridis6
Until now, we initialized the pipeline run on the Iridis6 HPC system
at the University of Southampton and developed a new, faster ingestion
method.

In this commit, I added a Slurm job file named 1_butler_ingest.slurm
for running the ingestion script. This script handles Butler
repository creation and ingestion of VIRCAM CDFS data and reference
catalogs on Iridis6.